requirements.txt
pytz==2021.3
python-dateutil==2.8.2
pillow==9.0.1
numpy==1.21.1
scipy==1.2.3
shapely==1.5.13
bottle==0.12.19
bintrees==2.2.0
jinja2==2.11.3
boto==2.49.0
